Move SealGenerationInfo tracking to the seal Access.
* Move SealGenerationInfo tracking to the seal Access.

The SealGenerationInfo is now kept track by a Seal's Access instead of by the
Config object. The access implementation now records the correct generation
number on seal wrapped values.

* Only store and read SealGenerationInfo if VAULT_ENABLE_SEAL_HA_BETA is true.

+++ vault/seal_wrapped_value.go
@@ -13,7 +13,7 @@ import (
 // transitoryGeneration is the Generation value used by SealWrappebValues for
 // entries that need to be upgraded from pre Vault 1.15.
 // Note that the value is 0, since that is the default value for BlobInfo.Generation.
-const transitoryGeneration = 0
+const transitoryGeneration = uint64(0)
 
 type SealWrappedValue struct {
 	value wrapping.MultiWrapValue
